Reworked C++ Clipper2Lib to be header only, templated by PointType, ZFillFunc and ClipperFlags (configuring reverse contour orientation).
